fix: make more or less stable movement of the entities for 2 players
Problem was that the movement for multiple players was laggy and removed the entity cell from the map. This solution fixes this problem, but as a drawback makes the wait the method call for the other soldiers from the same class.

This solution is definitely better than what we had before, though not ideal and needs refining
(Sometimes the creation of new soldiers might fail due to the re-render cells clear solution)

Some problems
- Re-render causes soldiers not to be created sometimes
- Soldiers from the same class will wait one each other to finish

Fixes:
- Graph gets now WorldMap class reference, useful for taking locked cell position
- new move soldier logic, algorithm: until positions are not the same do -> calculate the graph paths and change the blocks by one cell
- small fixes